Edit find_verified_device method in connection.rb
When reject_unauthorized_connection returned it was
blocking any connection including Turbo Streams as well.

So changing 'reject_unauthorized_connection' to 'nil'
allowed me to use Turbo Streams in other parts of the
app, whether it found a verified device or not.

This extra connection brings one advantage which is when
logged-in device closes the tab without signing-out user
status changes immediately. Then the partial gets updated
in real-time.

Now, realtime activity feature can be used as intended.